+++ routes/api/contactsRouter.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import contactsController from "../../controllers/contacts-controller.js";
+import contactsController from "../../controllers/contacts.js";
 
 import {isEmptyBody, isEmptyBodyPUT} from "../../middlewars/index.js";
 const router = express.Router()
